Forbid identical container name and version
If the configured repositories contain a container with the same name
and version twice do not add both to the index.

+++ northstar/src/runtime/island/init.rs
@@ -81,7 +81,7 @@ pub(super) fn init(
     env::set_current_dir("/").expect("Failed to set cwd to /");
 
     // UID / GID
-    setid(manifest.uid, manifest.gid);
+    setid(manifest.uid as u32, manifest.gid as u32);
 
     // Supplementary groups
     setgroups(groups);
